% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_download.R
\name{ABI_resample}
\alias{ABI_resample}
\title{Resample volume to ABI space}
\usage{
ABI_resample(
  source_volume,
  target_volume = NULL,
  target_space_name = NULL,
  xfm_transform = NULL,
  source_volume_resampled_file = NULL
)
}
\arguments{
\item{source_volume}{MINC file or vector you want to transform}

\item{target_volume}{MINC file or vector you want to resample to. If NULL, it is assumed you want to transform to ABI CCFv3 50 micron.}

\item{target_space_name}{Instead of specifying target_volume, you can provide one of several strings representing target volumes. Options include "adult gene expression" (i.e. 200um isotropic voxels), and "CCFv3 50um" (i.e. 50um isotropic voxels). If NULL, it is assumed you want to transform to ABI CCFv3 50 micron.}

\item{xfm_transform}{xfm file denoting the transform. Default NULL means no transformation}

\item{source_volume_resampled_file}{filepath to save resampled source_volume. If NULL (default), file is not saved.}
}
\value{
source_volume resampled to target space
}
\description{
Resample a volume to a target with optional transform using linear interpolation.
}
