% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{ABI_template_align}
\alias{ABI_template_align}
\title{Download and register to ABI 50um template}
\usage{
ABI_template_align(
  anatomy_template,
  anatomy_mask,
  allen_ccf3v_template = NULL,
  allen_ccf3v_mask = NULL,
  xfm_files_dir = tempdir(),
  keep_files = F
)
}
\arguments{
\item{anatomy_template}{minc data to move}

\item{anatomy_mask}{mask used in the registration}

\item{allen_ccf3v_template}{Path to the ABI template. If NULL, ABI 50um template is downloaded and stored to a temporary file. If path does not point to an existing file, the template is downloaded and written here. If path points to an existing file, the file is used as the target for registration}

\item{xfm_files_dir}{Directory to store the XFM transformation files. If NULL, a temporary directory is used.}

\item{keep_files}{If TRUE, pathnames to ABI template and XFM directories will be echoed. Its User's responsibility make sure these files are in a safe location.}
}
\value{
XFM transformation filepath
}
\description{
Downloads ABI 50um template and registers minc data to it
}
