% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{adult_gene_expression_analysis}
\alias{adult_gene_expression_analysis}
\title{Gene expression spatial enrichment}
\usage{
adult_gene_expression_analysis(
  anatomy_statistics,
  gene_expression_paths = NULL,
  target_threshold,
  contrast_threshold = NULL,
  symmetric_statistics = F,
  ABI_registration = list(anatomy_template = NULL, anatomy_mask = NULL,
    allen_ccf3v_template = NULL, allen_ccf3v_mask = NULL, anatomy_transform = NULL),
  gene_expression_analyis_options = list(interpolate_gene = F, reflect_gene = T,
    brain_mask = NULL),
  tmp_files_dir = NULL,
  parallel = NULL,
  conf_file = getOption("RMINC_BATCH_CONF")
)
}
\arguments{
\item{anatomy_statistics}{MINC file or vector denoting statistics at each voxel}

\item{gene_expression_paths}{Filenames or URLs pointing to gene expression data. If NULL, then genome-wide gene expression analysis is conducted and gene expression data is downloaded to temporary files.}

\item{target_threshold}{statistics greater than this value constitute the ROI}

\item{contrast_threshold}{statistics less than this value constitute the contract region. If NULL, then contrast region is assumed to be the whole brain.}

\item{symmetric_statistics}{Should the absolute value of statistics at each voxel be considered instead of the signed value?}

\item{ABI_registration}{arguments that can be supplied if you want to perform a registration prior to analysis. If you provide an XFM filepath as 'anatomy_transform', then the statistics will be transformed and resampled prior to gene expression analysis. If you provide 'anatomy_template' (filepath or MINC vector denoting the anatomy where statistics were conducted), 'anatomy_mask' (filepath or MINC vector denoting brain mask for anatomy), then the anatomy is registered to the ABI template and the resulting transformation applied to the statistics prior to gene expression computation. In this case, ABI template and mask are downloaded unless 'allen_ccf3v_template' and 'allen_ccf3v_mask' point to filepaths containing the respective information.}

\item{gene_expression_analyis_options}{options for gene expression analysis. 'interpolate_gene' flags whether nearest-neighbour imputation should be done to fill in missing gene expression voxels. If FALSE, then these voxels are ignored. 'reflect_gene' flags whether gene expression signal should be reflected across the sagittal midplane to fill in missing gene expression voxels in the opposite hemisphere. 'brain_mask' is a MINC file, RAW file, or vector identifying which voxels are in the brain. If NULL, the mask will be downloaded from the ABI.}

\item{tmp_files_dir}{location to store temporary files from the registration}

\item{parallel}{how many processors to run on (default=single processor). Specified as a two element vector, with the first element corresponding to the type of parallelization, and the second to the number of processors to use. For local running set the first element to "local" or "snowfall" for back-compatibility, anything else will be run with batchtools. Leaving this argument NULL runs sequentially.}

\item{conf_file}{A batchtools configuration file defaulting to \code{getOption("RMINC_BATCH_CONF")}}
}
\value{
gene expression enrichment expressed as fold-change
}
\description{
Conducts a gene expression spatial enrichment analysis for arbitrary anatomy statistics. Statistics above the target threshold constitute the target region-of-interest (ROI). Statistics below a contrast threshold constitute the contrast region. Gene expression spatial enrichment is calculated for any number of genes in the ABI gene expression atlas. Enrichments is computed using a fold-change measure: expression in target ROI divided by expression in contrast region. If contrast threshold is not suppled, then the contrast regions is assumed to be the whole brain.
}
