% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_download.R
\name{allen_gene_template_download}
\alias{allen_gene_template_download}
\title{Download ABI gene templates}
\usage{
allen_gene_template_download(
  age_dataset = c("adult", "P56", "P28", "P14", "P4", "E18.5", "E16.5", "E15.5",
    "E13.5", "E11.5"),
  outfile = NULL,
  labels = F,
  grid_annot = F,
  binarize = F
)
}
\arguments{
\item{age_dataset}{Choose the age of the template to download. Options are 'adult','P56','P28','P14','P4','E18.5','E16.5','E15.5','E13.5','E11.5'. Note 'adult' and 'P56' are aliases.}

\item{outfile}{Optional outfile to write}

\item{labels}{If true, then the labels are downloaded. Default is FALSE, so the template is downloaded. Warning: labels may have integers that can't be written to a MINC file.}

\item{grid_annot}{If true, then the grid annotations are downloaded. They are much lower resolution compared to template. Default is FALSE, so the template is downloaded. Warning: grid annotations may have integers that can't be written to a MINC file.}

\item{binarize}{If true, then values are binarized with threshold of 0.5}
}
\value{
template data as a 1D vector of class mincSingleDim
}
\description{
Downloads and reads the ABI gene template with an option to write to a mincfile
}
