% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{allen_s2p_template_download}
\alias{allen_s2p_template_download}
\title{Download ABI 50um template}
\usage{
allen_s2p_template_download(outfile = NULL, mask = F)
}
\arguments{
\item{outfile}{Optional outfile to write}

\item{mask}{If true, then the mask is downloaded. Default is FALSE, so the template is downloaded.}
}
\value{
template data as a 1D vector of class mincSingleDim
}
\description{
Downloads and reads the ABI 50um template with an option to write to a mincfile
}
