% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllStructureLabels.R
\docType{data}
\name{AllStructureLabels}
\alias{AllStructureLabels}
\title{All Label values, Structures, and Ontologies used in Allen Brain Atlas Annotations (25um resolution)}
\description{
All Label values, Structures, and Ontologies used in Allen Brain Atlas Annotations (25um resolution)
}
\references{
\url{http://api.brain-map.org/api/v2/data/query.csv?criteria=model::Structure,rma::criteria,[ontology_id$eq1],rma::options[order$eq\%27structures.graph_order\%27][num_rows$eqall]}
}
\author{
Darren Fernandes \email{fernadj2@gmail.com}
}
\keyword{AllStructureLabels}
