% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsify.R
\name{ANTsify}
\alias{ANTsify}
\title{Convert RMINC and ABIgeneRMINC data to ANTsR images}
\usage{
ANTsify(x)
}
\arguments{
\item{x}{RMINC or ABIgeneRMINC data}
}
\value{
ANTsR images
}
\description{
Convert RMINC and ABIgeneRMINC data to ANTsR images
}
\examples{
# Here is an example converting ABIgeneRMINC data to ANTsR image
nrp1Filename=system.file('extdata/Nrp1_P56_coronal_74272479_200um.zip',package="ABIgeneRMINC")
ABIvec = read.raw.gene(nrp1Filename)
antsvec1 = ANTsify(ABIvec)

# Here is an example converting MINC data to ANTsR image
statFilename=system.file('extdata/enrichment_stats.mnc',package="ABIgeneRMINC")
mincvec = mincGetVolume(statFilename)
antsvec2 = ANTsify(mincvec)

}
