% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{check_registration_quality_png}
\alias{check_registration_quality_png}
\title{Resample and check registration quality}
\usage{
check_registration_quality_png(
  source_file,
  target_file,
  source_transform,
  resampled_source_file = NULL,
  reg_pik = NULL
)
}
\arguments{
\item{source_file}{MINC file that is moved}

\item{target_file}{MINC file that is the target}

\item{source_transform}{XFM file mapping the source_file to the target_file}

\item{resampled_source_file}{MINC filename where the transformed sourcefile will be written. Temporary file is created if this is not supplied.}

\item{reg_pik}{PNG filename where the image will be written. Temporary file is created if this is not supplied.}
}
\value{
PNG filename where the image will be written
}
\description{
Creates a PNG for manual assessment of registration quality.
}
