% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/antsify.R
\name{deANTsify}
\alias{deANTsify}
\title{Turn ANTs image into ABIgeneRMINC data}
\usage{
deANTsify(x)
}
\arguments{
\item{x}{ANTsR images}
}
\value{
ABIgeneRMINC data
}
\description{
Turn ANTs image into ABIgeneRMINC data
}
\examples{
# Here is an example converting ABIgeneRMINC data to ANTsR image
nrp1Filename=system.file('extdata/Nrp1_P56_coronal_74272479_200um.zip',package="ABIgeneRMINC")
ABIvec = read.raw.gene(nrp1Filename)
antsvec = ANTsify(ABIvec)
y = deANTsify(antsvec)

}
