% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.all.allen.genes.R
\name{find.dev.gene.experiment}
\alias{find.dev.gene.experiment}
\title{List the Allen mouse development experiments associated with gene}
\usage{
find.dev.gene.experiment(geneAcronym)
}
\arguments{
\item{geneAcronym}{Acronym of gene to find experiments of}
}
\value{
Data frame with Experiment number, slices (coronal or sagittal), and URLs to download data
}
\description{
List the Allen mouse development experiments associated with gene
}
\examples{
# Find all mouse development experiments associated with Ntng1
find.dev.gene.experiment('Ntng1')
}
