% Generated by roxygen2 (4.0.1): do not edit by hand
\name{find.gene.experiment}
\alias{find.gene.experiment}
\title{List the Allen experiments assiciated with gene}
\usage{
find.gene.experiment(geneAcronym)
}
\arguments{
\item{geneAcronym}{Acronym of gene to find experiments of}
}
\value{
Data frame with Experiment number, slices (coronal or sagittal), and URLs to download data
}
\description{
List the Allen experiments assiciated with gene
}
\examples{
# Find all experiments associated with Itsn1
find.gene.experiment('Itsn1')
}

