% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_reshaping.R
\name{gene_pad_arr}
\alias{gene_pad_arr}
\title{Pad array to fit a different shape}
\usage{
gene_pad_arr(x, new.dims, side = c("left", "right", "both"), padval = 0)
}
\arguments{
\item{x}{Array to pad}

\item{new.dims}{new dimensions}

\item{side}{should padding be applied to the left, right, or both (ties are broken by the left)?}

\item{padval}{what value should be used for padding?}
}
\value{
array padded so it has dimensions of new.dims
}
\description{
Pad array to fit a different shape
}
\examples{
# read a gene expression file
filename=system.file('extdata/Pdyn_P56_coronal_71717084_200um.zip',package="ABIgeneRMINC")
gene.expression=read.raw.gene(filename)

# make it into an array 
gene.expression.array=array(gene.expression,rev(attr(gene.expression,'sizes')))
# synonymous with mincArray(gene.expression))

# dimensions of array
dim(gene.expression.array)
# 67 41 58

# pad the array
padded.gene.expression.array = gene_pad_arr(gene.expression.array,new.dims=c(100,99,98),side='both')

# dimensions of padded array
dim(padded.gene.expression.array)
# 100 99 98
}
