% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOrilla_run.R
\name{GOrilla_run}
\alias{GOrilla_run}
\title{Run Gene Ontology enrichment analysis on GOrilla}
\usage{
GOrilla_run(target.genes, background.genes, tmpdir = tempdir())
}
\arguments{
\item{target.genes}{genes of interest}

\item{background.genes}{pool of genes from which the target genes were selected}

\item{tmpdir}{path to directory where files will be stored}
}
\value{
data frame of enriched process ontologies, significance from hypergeometric test (with FDR), and associated genes
}
\description{
Run Gene Ontology enrichment analysis on GOrilla
}
\examples{
# Target list of genes
target.genes = readLines(system.file('extdata/GOtest/target.txt',package="ABIgeneRMINC"))
# Background genes
background.genes = readLines(system.file('extdata/GOtest/background.txt',package="ABIgeneRMINC"))

# Run GOrilla
go_results = GOrilla_run(target.genes, background.genes)
}
