% Generated by roxygen2 (4.0.1): do not edit by hand
\name{list.all.allen.genes}
\alias{list.all.allen.genes}
\title{List all genes in the Allen Gene Expression Atlas}
\usage{
list.all.allen.genes(outfile = NULL, parallel = NULL)
}
\arguments{
\item{outfile}{(OPTIONAL) Filename to output results}

\item{parallel}{Number of cores registered for parallel computation.
Default is NULL, computations performed serially.
If not NULL, requires the foreach & doMC pakages for parallel computations.}
}
\value{
Vector of all genes currently in the Allen Gene Expression Atlas
}
\description{
List all genes in the Allen Gene Expression Atlas
}
\examples{
# Find all genes in the Allen Gene Expression Atlas and
# Output results to file "genes.txt" and perform computations in parallel (4 cores).
\dontrun{
genes=list.all.allen.genes("genes.txt",parallel=4)
}
}

