% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.all.allen.genes.R
\name{list.all.dev.allen.genes}
\alias{list.all.dev.allen.genes}
\title{List all genes in the Allen Developmental Gene Expression Atlas}
\usage{
list.all.dev.allen.genes(outfile = NULL, parallel = NULL)
}
\arguments{
\item{outfile}{(OPTIONAL) Filename to output results}

\item{parallel}{Number of cores registered for parallel computation. 
Default is NULL, computations performed serially. 
If not NULL, requires the parallel package for computations.}
}
\value{
Vector of all genes currently in the Allen Developmental Gene Expression Atlas
}
\description{
List all genes in the Allen Developmental Gene Expression Atlas
}
\examples{
# Find all genes in the Allen Gene Expression Atlas and
# Output results to file "genes.txt" and perform computations in parallel (4 cores).
\dontrun{
genes=list.all.dev.allen.genes("genes.txt",parallel=4)
}
}
