% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{minc_ants_register}
\alias{minc_ants_register}
\title{Register two minc volumes using ANTs}
\usage{
minc_ants_register(
  source_volume,
  target_volume,
  source_mask_volume,
  target_mask_volume,
  xfm_file = NULL,
  verbose = T
)
}
\arguments{
\item{source_volume}{volume to move. Can be an RMINC object or filename.}

\item{target_volume}{reference volume. Can be an RMINC object or filename.}

\item{source_mask_volume}{volume for masking source_volume}

\item{target_mask_volume}{volume for masking target_volume}

\item{xfm_file}{location to save file. If NULL, file is saved in temporary directory.}

\item{verbose}{If true, ANTs messages are passed to console}
}
\value{
location of xfm transform file
}
\description{
Uses system ANTs installation to register minc volumes
}
