% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registration.R
\name{mincvol_pad_crop}
\alias{mincvol_pad_crop}
\title{Pad or crop volume}
\usage{
mincvol_pad_crop(infile, outfile, operation = c("pad", "crop"))
}
\arguments{
\item{infile}{input MINC file}

\item{outfile}{outfile MINC file}

\item{operation}{If 'pad', then 2mm is added to all sides. If crop, then 2mm is cropped from all sides.}
}
\description{
Pad or crop 2mm to volume on all sides
}
