% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read.raw.gene}
\alias{read.raw.gene}
\title{Read Raw Allen Gene Expression File}
\usage{
read.raw.gene(filename, labels = FALSE)
}
\arguments{
\item{filename}{Filename of Raw Gene Expression File}

\item{labels}{logical. If TRUE, then file is read as a string of integers}
}
\value{
1D vector of length 159326 containing Raw Gene Expression data
}
\description{
Read Raw Allen Gene Expression File
}
\examples{
# Download gene expression files from the Allen Brain Institute
# In this example, I downloaded Pdyn expression energy (http://api.brain-map.org/grid_data/download/71717084)
# Then Unzipped it, and renamed the energy.raw file to Pdyn_energy_experiment71717084.raw
# I included it in this library as an example

filename=system.file('extdata/Pdyn_energy_experiment71717084.raw',package="ABIgeneRMINC")
gene.expression=read.raw.gene(filename)
}

