% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.all.allen.genes.R
\name{find.gene.experiment}
\alias{find.gene.experiment}
\title{List the Allen adult mouse experiments associated with gene}
\usage{
find.gene.experiment(geneAcronym)
}
\arguments{
\item{geneAcronym}{Acronym of gene to find experiments of}
}
\value{
Data frame with Experiment number, slices (coronal or sagittal), and URLs to download data
}
\description{
List the Allen adult mouse experiments associated with gene
}
\examples{
# Find all experiments associated with Itsn1
find.gene.experiment('Itsn1')
}
